LayoutTests/fast/events/touch/gesture/gesture-click.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../js/resources/js-test-pre.js"></script>
</head>
<body onload="runTest();">
<div id="gesture">

<p id="description"></p>
<div id="console"></div>

<script>
var expectedMouseEvents = 4;
var mouseEventsReceived = 0;
var eventTypes = [ 'mousemove', 'mousedown', 'mouseup', 'click' ];

function gestureEventCallback(event)
{
    if (window.eventSender) {
        shouldBeEqualToString('event.type', eventTypes[mouseEventsReceived]);
        shouldBe('event.clientX', '10');
        shouldBe('event.clientY', '12');
        mouseEventsReceived++;
    } else {
        debug(event.type);
        debug(event.clientX);
        debug(event.clientY);
    }
}

// Because we may not have a gesture recognizer, we send a key press
// event to end the test without temporal flakiness.
function quitKeyToEndTest(event) {
    endTest();
}

function endTest()
{
    if (mouseEventsReceived < expectedMouseEvents) {
        debug('Gesture manager not implemented on this platform.');
    }
    // Bail.
    isSuccessfullyParsed();
    layoutTestController.notifyDone();
}

function runTest() {
    var div = document.getElementById('gesture');
    div.addEventListener("mousedown", gestureEventCallback, false);
    div.addEventListener("click", gestureEventCallback, false);
    div.addEventListener("mouseup", gestureEventCallback, false);
    div.addEventListener("mousemove", gestureEventCallback, false);
    document.addEventListener("keydown", quitKeyToEndTest, false);

    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    if (window.eventSender) {
        description("This tests basic gesture callbacks.");
        // A 'tap' gesture event should generate a sequence of mouse events.
        eventSender.gestureTap(10, 12);
        eventSender.leapForward(10);
        eventSender.keyDown(' ');
    } else {
        debug("This test requires DumpRenderTree.  Tap on the blue rect to log.")
    }
}
</script>
</body>
</html>
