LayoutTests/platform/chromium/fast/events/touch/gesture/gesture-click-expected.txt

This tests basic gesture callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.type is "mousemove"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.type is "mousedown"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.type is "mouseup"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.type is "click"
PASS event.clientX is 10
PASS event.clientY is 12
PASS successfullyParsed is true

TEST COMPLETE
